routes: fetch isSecretsSyncActivated at clients/counts route level
