+++ ui/app/routes/vault/cluster/clients/counts.ts
@@ -5,6 +5,7 @@
 
 import Route from '@ember/routing/route';
 import { service } from '@ember/service';
+import { DEBUG } from '@glimmer/env';
 import timestamp from 'core/utils/timestamp';
 import { getUnixTime } from 'date-fns';
 
@@ -28,6 +29,7 @@ export interface ClientsCountsRouteModel {
   versionHistory: ClientsVersionHistoryModel;
   activity?: ClientsActivityModel;
   activityError?: AdapterError;
+  isSecretsSyncActivated: boolean;
   startTimestamp: number;
   endTimestamp: number;
 }
@@ -37,6 +39,13 @@ interface ClientsCountsController extends Controller {
   end_time: number | undefined;
   ns: string | undefined;
   mountPath: string | undefined;
+}
+
+interface ActivationFlagsResponse {
+  data: {
+    activated: Array<string>;
+    unactivated: Array<string>;
+  };
 }
 
 export default class ClientsCountsRoute extends Route {
@@ -49,7 +58,10 @@ export default class ClientsCountsRoute extends Route {
     mountPath: { refreshModel: false, replace: true },
   };
 
-  async getActivity(start_time: number, end_time: number) {
+  async getActivity(
+    start_time: number,
+    end_time: number
+  ): Promise<[ClientsActivityModel | undefined, AdapterError | unknown] | [Record<string, never>, null]> {
     let activity, activityError;
     // if there is no billingStartTimestamp or selected start date initially we allow the user to manually choose a date
     // in that case bypass the query so that the user isn't stuck viewing the activity error
@@ -67,19 +79,44 @@ export default class ClientsCountsRoute extends Route {
     return [{}, null];
   }
 
+  async getActivatedFeatures() {
+    try {
+      const resp: ActivationFlagsResponse = await this.store
+        .adapterFor('application')
+        .ajax('/v1/sys/activation-flags', 'GET', { unauthenticated: true, namespace: null });
+      return resp.data?.activated;
+    } catch (error) {
+      if (DEBUG) console.error(error); // eslint-disable-line no-console
+      return [];
+    }
+  }
+
+  async isSecretsSyncActivated(activity: ClientsActivityModel | Record<string, never> | undefined) {
+    // if there are secrets, the feature is activated
+    if (activity && activity.total?.secret_syncs > 0) return true;
+
+    // otherwise check explicitly if the feature has been activated
+    const activatedFeatures = await this.getActivatedFeatures();
+    return activatedFeatures.includes('secret-sync');
+  }
+
   async model(params: ClientsCountsRouteParams) {
     const { config, versionHistory } = this.modelFor('vault.cluster.clients') as ClientsRouteModel;
     // we could potentially make an additional request to fetch the license and get the start date from there if the config request fails
     const startTimestamp = Number(params.start_time) || getUnixTime(config.billingStartTimestamp);
     const endTimestamp = Number(params.end_time) || getUnixTime(timestamp.now());
     const [activity, activityError] = await this.getActivity(startTimestamp, endTimestamp);
+
+    const isSecretsSyncActivated = await this.isSecretsSyncActivated(activity);
+
     return {
-      config,
-      versionHistory,
       activity,
       activityError,
-      startTimestamp,
+      config,
       endTimestamp,
+      isSecretsSyncActivated,
+      startTimestamp,
+      versionHistory,
     };
   }
