+++ ui/app/templates/vault/cluster/clients/counts.hbs
@@ -7,12 +7,13 @@
   @activity={{this.model.activity}}
   @activityError={{this.model.activityError}}
   @config={{this.model.config}}
-  @versionHistory={{this.model.versionHistory}}
-  @startTimestamp={{this.model.startTimestamp}}
   @endTimestamp={{this.model.endTimestamp}}
-  @namespace={{this.ns}}
+  @isSecretsSyncActivated={{this.model.isSecretsSyncActivated}}
   @mountPath={{this.mountPath}}
+  @namespace={{this.ns}}
   @onFilterChange={{this.updateQueryParams}}
+  @startTimestamp={{this.model.startTimestamp}}
+  @versionHistory={{this.model.versionHistory}}
 >
   {{outlet}}
 </Clients::Page::Counts>
