+++ ui/app/templates/vault/cluster/clients/counts/overview.hbs
@@ -5,6 +5,7 @@
 
 <Clients::Page::Overview
   @activity={{this.model.activity}}
+  @isSecretsSyncActivated={{this.model.isSecretsSyncActivated}}
   @versionHistory={{this.model.versionHistory}}
   @startTimestamp={{this.model.startTimestamp}}
   @endTimestamp={{this.model.endTimestamp}}
